requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Mail==0.9.1
email-validator==2.2.0
Werkzeug==3.0.3
APScheduler==3.10.4
icalendar==5.0.11
python-dotenv==1.0.0
PyMySQL==1.1.1
